+++ example/lib/pages/path_finding/human_path_finding.dart
@@ -1,8 +1,7 @@
 import 'package:bonfire/bonfire.dart';
 import 'package:example/pages/enemy/human_with_collision.dart';
 
-class HumanPathFinding extends HumanWithCollision
-    with MoveToPositionAlongThePath, TapGesture {
+class HumanPathFinding extends HumanWithCollision with PathFinding, TapGesture {
   HumanPathFinding({required Vector2 position}) : super(position: position) {
     setupMoveToPositionAlongThePath(
       pathLineStrokeWidth: 2,
@@ -14,7 +13,7 @@ class HumanPathFinding extends HumanWithCollision
 
   @override
   void onTapDownScreen(TapGestureEvent event) {
-    moveToPositionAlongThePath(event.worldPosition);
+    moveToPositionWithPathFinding(event.worldPosition);
     super.onTapDownScreen(event);
   }
 }